conversion fix

+++ vms/c++/predict.cpp
@@ -99,7 +99,7 @@ predict_compound_block_c_loop11:
         for (d = 0; d < num_proteins; d++)
         {
 #pragma HLS UNROLL
-            predictions[c][d] = (float)(out_buf[d] / (S_type)num_samples);
+            predictions[c][d] = out_buf[d] / (S_type)num_samples;
         }
     } // end compounds
 }
